+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 <details>
 <summary>View Changelog</summary>
 
-# 2.20.1
+# 2.21.0
 - Fixed ability stacks not rendering
 - Fixed rendering error when displaying a card with tribes outside of Act 1
 - Fixed ResetShields not re-setting lostShield to false under certain conditions
 - Fixed RemoveMaxEnergy not working as intended
 - Fixed custom AudioClips not loading correctly on Mac OSX
 - Added AllModificationInfos, AllModificationTypes, and modification syncing to SlotModificationManager
+- Added additional functionality to SlotModificationManager - Infos now store name and GUID
+- Added rulebook entry support for slot modifications - use extension method .SetRulebook() when adding your slot modification
+- Added RuleBookManager for adding custom rulebook entries (see wiki)
 - Added more shield-related extensions
 - Added some AbilityInfo-related extensions
 - Added ShieldManager.AllShieldAbilities and ShieldManager.AllShieldInfos for easier tracking of custom shield abilities